Fix Intro Err
2026.01.31
intro씬에서 플레이어 및 적이 생성되는 문제 해결

+++ Assets/02. Script/Common/Constants.cs
@@ -69,7 +69,7 @@ public class Constants
     public class SpawnZone
     {
         public Transform point;
-        public float radius;
+        public int radius;
     }
 
     [Serializable]
